+++ lib/src/index.ts
@@ -5,7 +5,7 @@ export * from './json';
 // Internals (for backend implementations)
 export * from './metadata';
 export * from './types';
-export * from './frontend_options';
+export * from './options';
 export { default as FrontendFunctions } from './frontend_functions';
 export { default as DecoratorFactory } from './decorator_factory';
 export { default as ObjectPropertySerializer } from './object_property_serializer';