+++ lib/src/json/index.ts
@@ -1,7 +1,7 @@
 import DecoratorFactory from '../decorator_factory';
 import FrontendFunctions from '../frontend_functions';
-import { DecoratorOptions, DeflateOptions, InflateOptions } from '../frontend_options';
 import { DEFAULT_PROJECTION, MetadataManager } from '../metadata';
+import { DecoratorOptions, DeflateOptions, InflateOptions } from '../options';
 import Constructor from "../types/constructor";
 import Util from '../types/util';
 import JsonType from "./json_type";
@@ -10,33 +10,33 @@ export const BACKEND_NAME = 'json';
 
 /**
  * Define serializer for given property or type
- * @param options _(optional)_ Custom type serializer and/or other options
+ * @param options Custom type serializer and/or other options
  * @returns Type/property decorator
  */
 export function Serialize<TSerialized extends JsonType, TOriginal>(
     options?: DecoratorOptions<TSerialized, TOriginal>
 ): (protoOrCtor: Object | Constructor<TOriginal>, propertyName?: string) => void {
-    return DecoratorFactory(BACKEND_NAME, options);
+    return DecoratorFactory<JsonType, TOriginal>(BACKEND_NAME, options);
 }
 
 /**
  * Serialize given serializable type instance to a JSON-compatible type
  * @param serializable Serializable type instance
- * @param options _(optional)_ Deflate options
+ * @param options Deflate options
  * @returns JSON-compatible type which can be safely passed to `JSON.serialize`
  */
 export function deflate<TOriginal>(
     serializable: TOriginal,
     options?: DeflateOptions<JsonType, TOriginal>
 ): JsonType {
-    return FrontendFunctions<JsonType>(BACKEND_NAME).deflate(serializable, options);
+    return FrontendFunctions(BACKEND_NAME).deflate<JsonType, TOriginal>(serializable, options);
 }
 
 /**
  * Construct/deserialize a serializable type instance from a JSON-compatible object
  * @param ctor Serializable type constructor function
  * @param serialized JSON-compatible object (e.g. returned from `JSON.parse`)
- * @param options _(optional)_ Inflate options
+ * @param options Inflate options
  * @returns Serializable type instance
  */
 export function inflate<TOriginal>(
@@ -44,7 +44,7 @@ export function inflate<TOriginal>(
     serialized: JsonType,
     options?: InflateOptions<JsonType, TOriginal>
 ): TOriginal {
-    return FrontendFunctions<JsonType>(BACKEND_NAME).inflate(ctor, serialized, options);
+    return FrontendFunctions(BACKEND_NAME).inflate<JsonType, TOriginal>(ctor, serialized, options);
 }
 
 // Types